Rename the package directory to 'deeplearning'
