+++ moa/src/test/java/moa/classifiers/deeplearning/CANDTest.java
@@ -18,7 +18,7 @@
 /**
  *
  */
-package moa.classifiers.neuralNetworks;
+package moa.classifiers.deeplearning;
 
 import junit.framework.Test;
 import junit.framework.TestSuite;
